the old source code used int and float for return value type of UITab…

+++ Millenium/Controllers/CollectionViewController/CollectionCategoryListViewController/CollecitonCategoryViewController.m
@@ -149,7 +149,7 @@
 
 #pragma mark - UItableView Datasource
 
--(int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return arrProduct.count;
 }
@@ -197,7 +197,7 @@
 
 #pragma mark - UItableView Delegate
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 90;
 }
